backend/services/auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
asyncpg==0.29.0
redis==5.0.1
python-dotenv==1.0.0
structlog==23.2.0
httpx==0.25.2